+++ js/main.js
@@ -1,4 +1,4 @@
-import { renderGallery } from './modules/render-gallery.js';
+import { renderGallery } from './modules/gallery.js';
 const COUNT_PICTURES = 25;
 
 renderGallery(COUNT_PICTURES);
