+++ js/modules/gallery.js
@@ -1,0 +1,8 @@
+import { getPictures } from './data.js';
+import { renderMiniPictures } from './mini-pictures.js';
+
+const renderGallery = (count) => {
+  renderMiniPictures(getPictures(count));
+};
+
+export { renderGallery };
