--- js/modules/render-gallery.js
@@ -1,8 +1,0 @@
-import { getPictureData } from './data.js';
-import { renderMiniPictures } from './render-mini-pictures.js';
-
-const renderGallery = (count) => {
-  renderMiniPictures(getPictureData(count));
-};
-
-export { renderGallery };
